SNOW-2371565: Enable async execution by default with auto-detection for blocking statements (#138)
Queries now use async mode unless explicitly disabled or the statement is PUT/GET (which requires blocking)
Add metrics for submit, inline poll, and wait phase timings for performance analysis.

Default behavior is submit, get query id, poll for results. Remaining TODO is to handle session refresh. This results in at least one additional round trip. For a user running in the same cloud region, this should be negligible (~5 ms). From my laptop hitting us-west-2, it adds another 180ms. I think this is the right tradeoff for us.

On the snowflake backend, this will cause all queries will be persisted. This is probably too expensive to keep this as the default long term due to the effect on hybrid tables latency and the load on FDB.